requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
arrow==1.3.0
attrs==25.4.0
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.6.4
dotenv==0.9.9
fitdecode==0.10.0
fitparse==1.2.0
flexcache==0.3
flexparser==0.4
frozenlist==1.8.0
gpxpy==1.6.2
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.1
Pint==0.24.4
platformdirs==4.3.8
propcache==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.4
rich==14.1.0
six==1.17.0
stravalib==2.4
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.22.0
